more refactorings
